#105 [Refactor] 폴더/파일 구조 개선, 리팩토링 - utils > keys.js 생성
