+++ src/apis/mypage.js
@@ -1,4 +1,4 @@
-import instance from './instance';
+import instance from './instance/instance';
 
 const getMypage = async () => {
   try {
